(#190) Make reviews subsystem return error codes
This code requires a bit more love as it is very clunky and brittle.

Probably rewrite it entirely.

+++ include/gcli/forges.h
@@ -406,7 +406,7 @@ struct gcli_forge_descriptor {
 
 	/**
 	 * Fetch MR/PR reviews including comments */
-	size_t (*get_reviews)(
+    int (*get_reviews)(
 		char const *owner,
 		char const *repo,
 		int pr,
