+++ include/gcli/github/review.h
@@ -36,9 +36,9 @@
 
 #include <gcli/review.h>
 
-size_t github_review_get_reviews(char const *owner,
-                                 char const *repo,
-                                 int pr,
-                                 gcli_pr_review **out);
+int github_review_get_reviews(char const *owner,
+                              char const *repo,
+                              int pr,
+                              gcli_pr_review **out);
 
 #endif /* GITHUB_REVIEW_H */
