+++ include/gcli/review.h
@@ -63,10 +63,10 @@ void gcli_review_reviews_free(gcli_pr_review *it, size_t size);
 
 void gcli_review_comments_free(gcli_pr_review_comment *it, size_t size);
 
-size_t gcli_review_get_reviews(char const *owner,
-                               char const *repo,
-                               int pr,
-                               gcli_pr_review **out);
+int gcli_review_get_reviews(char const *owner,
+                            char const *repo,
+                            int pr,
+                            gcli_pr_review **out);
 
 void gcli_review_print_review_table(gcli_pr_review const *headers,
                                     size_t headers_size);
